[RDCC-7025]buil failure fix
