+++ src/contractTest/java/uk/gov/hmcts/reform/professionalapi/provider/BulkCustomerDetailsProviderTest.java
@@ -5,49 +5,22 @@ import au.com.dius.pact.provider.junit5.PactVerificationContext;
 import au.com.dius.pact.provider.junit5.PactVerificationInvocationContextProvider;
 import au.com.dius.pact.provider.junitsupport.Provider;
 import au.com.dius.pact.provider.junitsupport.State;
-import au.com.dius.pact.provider.junitsupport.loader.PactBroker;
-import au.com.dius.pact.provider.junitsupport.loader.VersionSelector;
 import au.com.dius.pact.provider.spring.junit5.MockMvcTestTarget;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.TestTemplate;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Primary;
+import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.TestPropertySource;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import uk.gov.hmcts.reform.professionalapi.controller.feign.UserProfileFeignClient;
 import uk.gov.hmcts.reform.professionalapi.controller.internal.BulkCustomerDetailsInternalController;
 import uk.gov.hmcts.reform.professionalapi.controller.request.BulkCustomerRequest;
-import uk.gov.hmcts.reform.professionalapi.controller.request.validator.OrganisationCreationRequestValidator;
-import uk.gov.hmcts.reform.professionalapi.controller.request.validator.PaymentAccountValidator;
-import uk.gov.hmcts.reform.professionalapi.controller.request.validator.ProfessionalUserReqValidator;
-import uk.gov.hmcts.reform.professionalapi.controller.request.validator.UpdateOrganisationRequestValidator;
-import uk.gov.hmcts.reform.professionalapi.controller.request.validator.UserProfileUpdateRequestValidator;
-import uk.gov.hmcts.reform.professionalapi.controller.request.validator.impl.OrganisationIdentifierValidatorImpl;
 import uk.gov.hmcts.reform.professionalapi.controller.response.BulkCustomerOrganisationsDetailResponse;
 import uk.gov.hmcts.reform.professionalapi.domain.BulkCustomerDetails;
 import uk.gov.hmcts.reform.professionalapi.domain.Organisation;
 import uk.gov.hmcts.reform.professionalapi.domain.OrganisationStatus;
 import uk.gov.hmcts.reform.professionalapi.repository.BulkCustomerDetailsRepository;
-import uk.gov.hmcts.reform.professionalapi.repository.ContactInformationRepository;
-import uk.gov.hmcts.reform.professionalapi.repository.DxAddressRepository;
-import uk.gov.hmcts.reform.professionalapi.repository.IdamRepository;
-import uk.gov.hmcts.reform.professionalapi.repository.OrganisationMfaStatusRepository;
-import uk.gov.hmcts.reform.professionalapi.repository.OrganisationRepository;
-import uk.gov.hmcts.reform.professionalapi.repository.PaymentAccountRepository;
-import uk.gov.hmcts.reform.professionalapi.repository.PrdEnumRepository;
-import uk.gov.hmcts.reform.professionalapi.repository.ProfessionalUserRepository;
-import uk.gov.hmcts.reform.professionalapi.service.MfaStatusService;
-import uk.gov.hmcts.reform.professionalapi.service.PaymentAccountService;
-import uk.gov.hmcts.reform.professionalapi.service.PrdEnumService;
-import uk.gov.hmcts.reform.professionalapi.service.ProfessionalUserService;
-import uk.gov.hmcts.reform.professionalapi.service.UserAccountMapService;
-import uk.gov.hmcts.reform.professionalapi.service.UserAttributeService;
 import uk.gov.hmcts.reform.professionalapi.service.impl.OrganisationServiceImpl;
 
 import java.util.UUID;
@@ -57,15 +30,10 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.powermock.api.mockito.PowerMockito.when;
 
 
-@ExtendWith(SpringExtension.class)
-@PactBroker(scheme = "${PACT_BROKER_SCHEME:http}",
-        host = "${PACT_BROKER_URL:localhost}",
-        port = "${PACT_BROKER_PORT:80}", consumerVersionSelectors = {
-        @VersionSelector(tag = "master")})
-@ContextConfiguration(classes = {BulkCustomerDetailsInternalController.class})
 @Provider("referenceData_bulkCustomerDetails")
-@TestPropertySource(locations = "/application-contract.yaml")
-public class BulkCustomerDetailsProviderTest {
+@ContextConfiguration(classes = {BulkCustomerDetailsInternalController.class})
+@Import(BulkCustomerDetailsProviderTestConfiguration.class)
+public class BulkCustomerDetailsProviderTest extends MockMvcProviderTest {
 
     public static final String ORG_NAME = "Org-Name";
     public static final String SRA_ID = "sra-id";
@@ -78,89 +46,15 @@ public class BulkCustomerDetailsProviderTest {
     BulkCustomerDetailsInternalController bulkCustomerDetailsInternalController;
 
     @MockBean
-    BulkCustomerDetailsRepository bulkCustomerDetailsRepository;
-
-    @MockBean
-    UserAccountMapService userAccountMapService;
-
-    @MockBean
-    BulkCustomerRequest bulkCustomerRequest;
-
-
-    @MockBean
-    OrganisationCreationRequestValidator organisationCreationRequestValidator;
-
-    @MockBean
-    ProfessionalUserReqValidator professionalUserReqValidator;
-
-    @MockBean
-    OrganisationIdentifierValidatorImpl organisationIdentifierValidator;
-
-    @MockBean
     OrganisationServiceImpl organisationService;
 
     @MockBean
-    UpdateOrganisationRequestValidator updateOrganisationRequestValidator;
-
-    @MockBean
-    PrdEnumService prdEnumService;
-
-    @MockBean
-    protected ProfessionalUserService professionalUserService;
-
-    @MockBean
-    ProfessionalUserRepository professionalUserRepository;
-
-    @MockBean
-    protected UserProfileUpdateRequestValidator userProfileUpdateRequestValidator;
-
-    @MockBean
-    protected PaymentAccountService paymentAccountService;
-
-    @MockBean
-    OrganisationRepository organisationRepository;
+    BulkCustomerDetailsRepository bulkCustomerDetailsRepository;
 
 
-    @MockBean
-    PaymentAccountRepository paymentAccountRepository;
-
-    @MockBean
-    DxAddressRepository dxAddressRepository;
-
-    @MockBean
-    ContactInformationRepository contactInformationRepository;
-
-    @MockBean
-    PrdEnumRepository prdEnumRepository;
-
-    @MockBean
-    UserAttributeService userAttributeService;
-
-    @MockBean
-    OrganisationMfaStatusRepository organisationMfaStatusRepository;
-
-    @MockBean
-    PaymentAccountValidator paymentAccountValidator;
-
-    @MockBean
-    UserProfileFeignClient userProfileFeignClient;
-
-    @MockBean
-    MfaStatusService mfaStatusService;
-
-    @MockBean
-    IdamRepository idamRepository;
-
-    @Value("${resendInviteEnabled}")
-    private boolean resendInviteEnabled;
-
-    @MockBean
-    BulkCustomerOrganisationsDetailResponse response;
-
-    @Bean
-    @Primary
-    protected OrganisationServiceImpl organisationService() {
-        return new OrganisationServiceImpl();
+    @Override
+    void setController() {
+        testTarget.setControllers(bulkCustomerDetailsInternalController);
     }
 
     @TestTemplate
